feat(desserbee-web): add useTag hook
