+++ apps/desserbee-web/src/app/[lang]/(user)/(search)/map/_components/PreferenceTags/index.tsx
@@ -6,7 +6,7 @@ import {
 
 import { Tag } from '@repo/design-system/components/Tag';
 import { cn } from '@repo/ui/lib/utils';
-import { useState } from 'react';
+import { useTag } from '../../../_hooks/useTag';
 
 interface PreferenceTagsProps {
   categories: string[];
@@ -17,31 +17,12 @@ export function PreferenceTags({
   categories,
   userPreferences,
 }: PreferenceTagsProps) {
-  const [selectedCategories, setSelectedCategories] = useState<Set<string>>(
-    new Set(),
-  );
-  const [isMyPreferSelected, setIsMyPreferSelected] = useState(false);
-
-  const handleTagClick = (category: string) => {
-    setSelectedCategories((prev) => {
-      const newSet = new Set(prev);
-      if (newSet.has(category)) {
-        newSet.delete(category);
-      } else {
-        newSet.add(category);
-      }
-      return newSet;
-    });
-    setIsMyPreferSelected(false);
-  };
-
-  const handleMyPreferenceTagClick = () => {
-    setSelectedCategories((prev) => {
-      const newSet = new Set([...userPreferences]);
-      return newSet;
-    });
-    setIsMyPreferSelected(true);
-  };
+  const {
+    isMyPreferSelected,
+    selectedCategories,
+    handleMyPreferenceTagClick,
+    handleTagClick,
+  } = useTag();
 
   return (
     <Carousel
@@ -59,12 +40,12 @@ export function PreferenceTags({
               'text-lg font-medium select-none text-nowrap text-[#DE8332]',
               isMyPreferSelected && 'text-white bg-[#DE8332]',
             )}
-            onClick={handleMyPreferenceTagClick}
+            onClick={() => handleMyPreferenceTagClick(userPreferences)}
           >
             내취향
           </Tag>
         </div>
-        {categories.map((category, index) => (
+        {categories.map((category) => (
           <CarouselItem
             key={category}
             className={`pl-2 basis-1/${categories.length} text-nowrap`}
